feat: 대시보드 레이아웃에서 프로그램 추천 섹션 리팩토링
- `RecommendedProgramSwiper` 컴포넌트를 `RecommendedProgramSection`으로 분리하여 코드 구조 개선
- `RecommendedProgramSection`에서 챌린지 데이터를 조회하고 추천 프로그램을 표시하도록 변경
- `RecommendedProgramSwiper`에 프로그램 리스트를 props로 전달하여 재사용성 향상
